Automated generation of test cases and revised tests
Instead of hand crafting test cases an example.sh script will loop through the input files in the example folder and generate the various outputs.

Unit tests are then used to validate that these match.  This allows for not having to hand craft thet test cases as this is revised and instead able to validate them as changed before committing to git.